refactor(prj): separate logger to core module
