+++ core/go.mod
@@ -14,6 +14,7 @@ require (
 	go.uber.org/zap v1.27.0
 	golang.org/x/net v0.25.0
 	golang.org/x/sync v0.7.0
+	gopkg.in/natefinch/lumberjack.v2 v2.2.1
 )
 
 require (
